+++ tsup.config.ts
@@ -8,6 +8,6 @@ export default defineConfig({
   sourcemap: true,
   clean: true,
   treeshake: true,
-  minify: false,
+  minify: true,
   target: 'es2022',
 })